feat: add embedding API with data editor, shared library build, and double-precision axis limits
- Add EmbedSurface class for headless figure rendering to memory buffers
- Add DataEditor overlay for interactive series data editing with undo/redo
- Add SPECTRA_BUILD_EMBED_SHARED CMake option to build libspectra_embed.so for Python ctypes/C FFI
- Add easy_embed.hpp API for simplified embedding workflow
- Add easy_embed_demo.cpp and qt_embed_demo.cpp examples (Qt demo requires SPECTRA_HAS_QT6)
- Change

+++ include/spectra/axes.hpp
@@ -34,8 +34,8 @@ struct AxisStyle
 
 struct AxisLimits
 {
-    float min = 0.0f;
-    float max = 1.0f;
+    double min = 0.0;
+    double max = 1.0;
 };
 
 struct TickResult
@@ -125,8 +125,8 @@ class Axes : public AxesBase
     BarSeries&       bar(std::span<const float> positions, std::span<const float> heights);
 
     // Axis configuration
-    void xlim(float min, float max);
-    void ylim(float min, float max);
+    void xlim(double min, double max);
+    void ylim(double min, double max);
     void title(const std::string& t);
     void xlabel(const std::string& lbl);
     void ylabel(const std::string& lbl);
